Move to utility class.
